web: remove redundant pipeline info from job
The Job record contains both a pipeline and team name, and prior to this
change, also a `PipelineIdentifier` that contains this same information.
This makes it awkward to decode jobs, as we provide the pipeline
identifier prior to decoding. When we decode the "list all jobs"
response, we pass in a meaningless identifier (since we can't possibly
know it prior to decoding).

We also don't appear to use the `Job.pipeline` field anywhere, so we can
safely remove it.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>